ADD label in BMRC config to use local software
Created a label 'local_software' to set the container
directive to null so that locally installed software
is used on the BMRC for processes with this label.
This label is set in the CONCAT_FASTQ and GUNZIP
processes.

Also added the Ubuntu container for CONCAT_FASTQ
process for use on laptop