Add restrander config
Added restrander config for cDNA sequencing. LSK114 kit was used
in BD2 pilot.

Slack message from Adam:
"SSP: TTT CTG TTG GTG CTG ATA TTG CTmG mGmG
VNP:ACTTGCCTGTCGCTCTATCTTCTTTTTTTTTTTTTTTTTTTTVN
The VNP and SSP primer sequences... for the VNP, the VN at the end
are degenerate bases so could be any base. Both were copied directly from the ONT documentation.
Not sure if it will be happy with either the VN or the mG's, but if you remove them and the m's from the mG
(it's just because theyre modified bases)
Also maybe need to take off the T's and just put how many T's there were in the length category in the json"

The sequences provided by Adam seem to match up with the config file
DCS-LSK114.json on restrander github